+++ guillotina/db/storages/pg.py
@@ -578,6 +578,29 @@ class PGConnectionManager:
         await self._initialize_tid_statements()
 
 
+class TransactionConnectionContextManager:
+    """
+    Connection manager to either use reserved
+    transaction connection or use a connection
+    from pool
+    """
+
+    def __init__(self, storage, txn):
+        self.storage = storage
+        self.txn = txn
+        self.connection = None
+
+    async def __aenter__(self):
+        if self.txn.connection_reserved:
+            return self.txn._db_conn
+        self.connection = await self.storage.pool.acquire(timeout=self.storage._conn_acquire_timeout)
+        return self.connection
+
+    async def __aexit__(self, *exc):
+        if self.connection is not None:
+            await self.storage.pool.release(self.connection)
+
+
 @implementer(IPostgresStorage)
 class PostgresqlStorage(BaseStorage):
     """Storage to a relational database, based on invalidation polling"""
@@ -645,7 +668,6 @@ class PostgresqlStorage(BaseStorage):
         read_only=False,
         name=None,
         pool_size=13,
-        transaction_strategy="resolve_readcommitted",
         conn_acquire_timeout=20,
         db_schema="public",
         store_json=True,
@@ -655,7 +677,7 @@ class PostgresqlStorage(BaseStorage):
         autovacuum=True,
         **options,
     ):
-        super(PostgresqlStorage, self).__init__(read_only, transaction_strategy=transaction_strategy)
+        super(PostgresqlStorage, self).__init__(read_only)
         self._dsn = dsn
         self._pool_size = pool_size
         self._partition_class = partition
@@ -927,10 +949,12 @@ WHERE tablename = '{}' AND indexname = '{}_parent_id_id_key';
         if self._connection_manager._vacuum is not None:
             await self._connection_manager._vacuum.add_to_queue(oid, self.objects_table_name)
 
+    def acquire(self, txn):
+        return TransactionConnectionContextManager(self, txn)
+
     async def delete(self, txn, oid):
-        conn = await txn.get_connection()
-        sql = self._sql.get("TRASH_PARENT_ID", self.objects_table_name)
-        async with watch_lock(txn._lock, "delete_object"):
+        sql = self._sql.get("TRASH_PARENT_ID", self._objects_table_name)
+        async with self.acquire(txn) as conn:
             # for delete, we reassign the parent id and delete in the vacuum task
             with watch("delete_object"):
                 await conn.execute(sql, oid)
@@ -949,26 +973,27 @@ WHERE tablename = '{}' AND indexname = '{}_parent_id_id_key';
 
     @restart_conn_on_exception
     async def get_next_tid(self, txn):
-        with watch("next_tid"):
-            async with self.pool.acquire(timeout=self._conn_acquire_timeout) as conn:
+        async with self.pool.acquire(timeout=self._conn_acquire_timeout) as conn:
+            with watch("next_tid"):
                 return await conn.fetchval(self._next_tid_sql.format(schema=self._db_schema))
 
     @restart_conn_on_exception
     async def get_current_tid(self, txn):
-        with watch("current_tid"):
-            async with self.pool.acquire(timeout=self._conn_acquire_timeout) as conn:
+        async with self.pool.acquire(timeout=self._conn_acquire_timeout) as conn:
+            with watch("current_tid"):
                 return await conn.fetchval(self._max_tid_sql.format(schema=self._db_schema))
 
-    async def get_one_row(self, txn, sql, *args):
-        conn = await txn.get_connection()
+    async def get_one_row(self, txn, sql, *args, prepare=False, metric="get_one_row"):
         # Helper function to provide easy adaptation to cockroach
-        return await conn.fetchrow(sql, *args)
+        async with self.acquire(txn) as conn:
+            with watch(metric):
+                return await conn.fetchrow(sql, *args)
 
     def _db_transaction_factory(self, txn):
         # make sure asycpg knows this is a new transaction
         if txn._db_conn._con is not None:
             txn._db_conn._con._top_xact = None
-        return txn._db_conn.transaction(readonly=txn._manager._storage._read_only)
+        return txn._db_conn.transaction(readonly=self._read_only)
 
     @restart_conn_on_exception
     async def _async_db_transaction_factory(self, txn):
@@ -1014,29 +1039,28 @@ WHERE tablename = '{}' AND indexname = '{}_parent_id_id_key';
                 return await self.start_transaction(txn, retries + 1)
 
     async def get_conflicts(self, txn):
-        async with watch_lock(self.lock, "shared_conflicts"):
-            if len(txn.modified) == 0:
-                return []
-            # use storage lock instead of transaction lock
-            if len(txn.modified) < 1000:
-                # if it's too large, we're not going to check on object ids
-                modified_oids = [k for k in txn.modified.keys()]
-                sql = self._sql.get("TXN_CONFLICTS_ON_OIDS", self.objects_table_name)
-                with watch("get_conflicts_oids"):
-                    async with self.pool.acquire() as conn:
-                        return await conn.fetch(sql, txn._tid, modified_oids)
-            else:
-                sql = self._sql.get("TXN_CONFLICTS", self.objects_table_name)
-                with watch("get_conflicts"):
-                    async with self.pool.acquire() as conn:
-                        return await conn.fetch(sql, txn._tid)
+        if len(txn.modified) == 0:
+            return []
+        # use storage lock instead of transaction lock
+        if len(txn.modified) < 1000:
+            # if it's too large, we're not going to check on object ids
+            modified_oids = [k for k in txn.modified.keys()]
+            sql = self._sql.get("TXN_CONFLICTS_ON_OIDS", self._objects_table_name)
+            with watch("get_conflicts_oids"):
+                async with self.pool.acquire() as conn:
+                    return await conn.fetch(sql, txn._tid, modified_oids)
+        else:
+            sql = self._sql.get("TXN_CONFLICTS", self._objects_table_name)
+            with watch("get_conflicts"):
+                async with self.pool.acquire() as conn:
+                    return await conn.fetch(sql, txn._tid)
 
     async def commit(self, transaction):
         async with watch_lock(transaction._lock, "commit_txn"):
             if transaction._db_txn is not None:
                 with watch("commit_txn"):
                     await transaction._db_txn.commit()
-            elif self._transaction_strategy not in ("none", "tidonly") and not transaction._skip_commit:
+            elif not transaction._skip_commit:
                 log.warning("Do not have db transaction to commit")
             return transaction._tid
 
@@ -1055,77 +1079,67 @@ WHERE tablename = '{}' AND indexname = '{}_parent_id_id_key';
 
     # Introspection
     async def get_page_of_keys(self, txn, oid, page=1, page_size=1000):
-        conn = await txn.get_connection()
         keys = []
-        sql = self._sql.get("BATCHED_GET_CHILDREN_KEYS", self.objects_table_name)
-        with watch("page_of_keys"):
-            for record in await conn.fetch(sql, oid, page_size, (page - 1) * page_size):
-                keys.append(record["id"])
+        sql = self._sql.get("BATCHED_GET_CHILDREN_KEYS", self._objects_table_name)
+        async with self.acquire(txn) as conn:
+            with watch("page_of_keys"):
+                for record in await conn.fetch(sql, oid, page_size, (page - 1) * page_size):
+                    keys.append(record["id"])
         return keys
 
     async def keys(self, txn, oid):
-        conn = await txn.get_connection()
-        sql = self._sql.get("GET_CHILDREN_KEYS", self.objects_table_name)
-        async with watch_lock(txn._lock, "keys"):
+        sql = self._sql.get("GET_CHILDREN_KEYS", self._objects_table_name)
+        async with self.acquire(txn) as conn:
             with watch("keys"):
                 result = await conn.fetch(sql, oid)
         return result
 
     async def get_child(self, txn, parent_oid, id):
-        sql = self._sql.get("GET_CHILD", self.objects_table_name)
-        async with watch_lock(txn._lock, "get_child"):
-            with watch("get_child"):
-                result = await self.get_one_row(txn, sql, parent_oid, id)
+        sql = self._sql.get("GET_CHILD", self._objects_table_name)
+        result = await self.get_one_row(txn, sql, parent_oid, id, metric="get_child")
         return result
 
     async def get_children(self, txn, parent_oid, ids):
-        conn = await txn.get_connection()
-        sql = self._sql.get("GET_CHILDREN_BATCH", self.objects_table_name)
-        async with watch_lock(txn._lock, "get_children"):
+        sql = self._sql.get("GET_CHILDREN_BATCH", self._objects_table_name)
+        async with self.acquire(txn) as conn:
             with watch("get_children"):
                 return await conn.fetch(sql, parent_oid, ids)
 
     async def has_key(self, txn, parent_oid, id):
-        sql = self._sql.get("EXIST_CHILD", self.objects_table_name)
-        async with watch_lock(txn._lock, "has_key"):
-            with watch("has_key"):
-                result = await self.get_one_row(txn, sql, parent_oid, id)
+        sql = self._sql.get("EXIST_CHILD", self._objects_table_name)
+        result = await self.get_one_row(txn, sql, parent_oid, id, metric="has_key")
         if result is None:
             return False
         else:
             return True
 
     async def len(self, txn, oid):
-        conn = await txn.get_connection()
-        sql = self._sql.get("NUM_CHILDREN", self.objects_table_name)
-        async with watch_lock(txn._lock, "num_children"):
+        sql = self._sql.get("NUM_CHILDREN", self._objects_table_name)
+        async with self.acquire(txn) as conn:
             with watch("num_children"):
                 result = await conn.fetchval(sql, oid)
         return result
 
     async def items(self, txn, oid):
-        conn = await txn.get_connection()
-        sql = self._sql.get("GET_CHILDREN", self.objects_table_name)
-        with watch("items"):
-            # not going to be accurate measure but will tell you if it is abused
-            async for record in conn.cursor(sql, oid):
-                # locks are dangerous in cursors since comsuming code might do
-                # sub-queries and they you end up with a deadlock
-                yield record
+        sql = self._sql.get("GET_CHILDREN", self._objects_table_name)
+        async with self.acquire(txn) as conn:
+            with watch("items"):
+                # not going to be accurate measure but will tell you if it is abused
+                async for record in conn.cursor(sql, oid):
+                    # locks are dangerous in cursors since comsuming code might do
+                    # sub-queries and they you end up with a deadlock
+                    yield record
 
     async def get_annotation(self, txn, oid, id):
-        sql = self._sql.get("GET_ANNOTATION", self.objects_table_name)
-        async with watch_lock(txn._lock, "load_annotation"):
-            with watch("load_annotation"):
-                result = await self.get_one_row(txn, sql, oid, id)
-            if result is not None and result["parent_id"] == TRASHED_ID:
-                result = None
+        sql = self._sql.get("GET_ANNOTATION", self._objects_table_name)
+        result = await self.get_one_row(txn, sql, oid, id, metric="load_annotation")
+        if result is not None and result["parent_id"] == TRASHED_ID:
+            result = None
         return result
 
     async def get_annotation_keys(self, txn, oid):
-        conn = await txn.get_connection()
-        sql = self._sql.get("GET_ANNOTATIONS_KEYS", self.objects_table_name)
-        async with watch_lock(txn._lock, "load_annotation_keys"):
+        sql = self._sql.get("GET_ANNOTATIONS_KEYS", self._objects_table_name)
+        async with self.acquire(txn) as conn:
             with watch("load_annotation_keys"):
                 result = await conn.fetch(sql, oid)
         items = []
@@ -1135,14 +1149,12 @@ WHERE tablename = '{}' AND indexname = '{}_parent_id_id_key';
         return items
 
     async def write_blob_chunk(self, txn, bid, oid, chunk_index, data):
-        sql = self._sql.get("HAS_OBJECT", self.objects_table_name)
-        async with watch_lock(txn._lock, "has_object"):
-            result = await self.get_one_row(txn, sql, oid)
+        sql = self._sql.get("HAS_OBJECT", self._objects_table_name)
+        result = await self.get_one_row(txn, sql, oid, metric="has_object")
         if result is None:
             # check if we have a referenced ob, could be new and not in db yet.
             # if so, create a stub for it here...
-            conn = await txn.get_connection()
-            async with watch_lock(txn._lock, "store_blob_stub"):
+            async with self.acquire(txn) as conn:
                 with watch("store_blob_stub"):
                     await conn.execute(
                         f"""INSERT INTO {self.objects_table_name}
@@ -1150,62 +1162,54 @@ WHERE tablename = '{}' AND indexname = '{}_parent_id_id_key';
     VALUES ($1::varchar({MAX_UID_LENGTH}), -1, 0, 0, TRUE, 'stub')""",
                         oid,
                     )
-        conn = await txn.get_connection()
-        sql = self._sql.get("INSERT_BLOB_CHUNK", self.blobs_table_name)
-        async with watch_lock(txn._lock, "store_blob_chunk"):
+        sql = self._sql.get("INSERT_BLOB_CHUNK", self._blobs_table_name)
+        async with self.acquire(txn) as conn:
             with watch("store_blob_chunk"):
                 return await conn.execute(sql, bid, oid, chunk_index, data)
 
     async def read_blob_chunk(self, txn, bid, chunk=0):
-        sql = self._sql.get("READ_BLOB_CHUNK", self.blobs_table_name)
-        async with watch_lock(txn._lock, "load_blob_chunk"):
-            with watch("load_blob_chunk"):
-                return await self.get_one_row(txn, sql, bid, chunk)
+        sql = self._sql.get("READ_BLOB_CHUNK", self._blobs_table_name)
+        return await self.get_one_row(txn, sql, bid, chunk, metric="load_blob_chunk")
 
     async def read_blob_chunks(self, txn, bid):
-        conn = await txn.get_connection()
-        with watch("read_blob_chunks"):
-            # again, not accurate through an iterator
-            async for record in conn.cursor(bid):
-                # locks are dangerous in cursors since comsuming code might do
-                # sub-queries and they you end up with a deadlock
-                yield record
+        async with self.acquire(txn) as conn:
+            with watch("read_blob_chunks"):
+                # again, not accurate through an iterator
+                async for record in conn.cursor(bid):
+                    # locks are dangerous in cursors since comsuming code might do
+                    # sub-queries and they you end up with a deadlock
+                    yield record
 
     async def del_blob(self, txn, bid):
-        conn = await txn.get_connection()
-        sql = self._sql.get("DELETE_BLOB", self.blobs_table_name)
-        async with watch_lock(txn._lock, "delete_blob_chunk"):
+        sql = self._sql.get("DELETE_BLOB", self._blobs_table_name)
+        async with self.acquire(txn) as conn:
             with watch("delete_blob_chunk"):
                 await conn.execute(sql, bid)
 
     async def get_total_number_of_objects(self, txn):
-        conn = await txn.get_connection()
-        sql = self._sql.get("NUM_ROWS", self.objects_table_name)
-        async with watch_lock(txn._lock, "total_objects"):
+        sql = self._sql.get("NUM_ROWS", self._objects_table_name)
+        async with self.acquire(txn) as conn:
             with watch("total_objects"):
                 result = await conn.fetchval(sql)
         return result
 
     async def get_total_number_of_resources(self, txn):
-        conn = await txn.get_connection()
-        sql = self._sql.get("NUM_RESOURCES", self.objects_table_name)
-        async with watch_lock(txn._lock, "total_resources"):
+        sql = self._sql.get("NUM_RESOURCES", self._objects_table_name)
+        async with self.acquire(txn) as conn:
             with watch("total_resources"):
                 result = await conn.fetchval(sql)
         return result
 
     async def get_total_resources_of_type(self, txn, type_):
-        conn = await txn.get_connection()
-        sql = self._sql.get("NUM_RESOURCES_BY_TYPE", self.objects_table_name)
-        async with watch_lock(txn._lock, "total_objects_by_type"):
+        sql = self._sql.get("NUM_RESOURCES_BY_TYPE", self._objects_table_name)
+        async with self.acquire(txn) as conn:
             with watch("total_objects_by_type"):
                 result = await conn.fetchval(sql, type_)
         return result
 
     # Massive treatment without security
     async def _get_page_resources_of_type(self, txn, type_, page, page_size):
-        conn = await txn.get_connection()
-        async with txn._lock:
+        async with self.acquire(txn) as conn:
             keys = []
             sql = self._sql.get("RESOURCES_BY_TYPE", self.objects_table_name)
             for record in await conn.fetch(sql, type_, page_size, (page - 1) * page_size):